requirements.txt
asn1crypto==0.24.0
cryptography==2.1.4
django-countries==6.1.2
django-widget-tweaks==1.4.8
enum34==1.1.6
gyp==0.1
idna==2.6
ipaddress==1.0.17
keyring==10.6.0
keyrings.alt==3.0
Pillow==6.2.2
pycrypto==2.6.1
pygobject==3.26.1
pyxdg==0.25
SecretStorage==2.3.1
six==1.11.0
